Refactor Action to use @actions/core and @actions/github instead of deprecated actions-toolkit
