+++ index.js
@@ -1,50 +1,3 @@
-const { Toolkit } = require('actions-toolkit')
-const getConfig = require('./utils/config')
-const matcher = require('matcher')
+const action = require('./action')
 
-const CONFIG_FILENAME = 'pr-labeler.yml'
-const defaults = {
-  feature: ['feature/*', 'feat/*'],
-  fix: 'fix/*',
-  chore: 'chore/*'
-}
-
-Toolkit.run(
-  async tools => {
-    const repoInfo = {
-      owner: tools.context.payload.repository.owner.login,
-      repo: tools.context.payload.repository.name
-    }
-    const ref = tools.context.payload.pull_request.head.ref
-    const config = {
-      ...defaults,
-      ...(await getConfig(tools.github, CONFIG_FILENAME, repoInfo, ref))
-    }
-
-    const labelsToAdd = Object.entries(config).reduce(
-      (labels, [label, patterns]) => {
-        if (
-          Array.isArray(patterns)
-            ? patterns.some(pattern => matcher.isMatch(ref, pattern))
-            : matcher.isMatch(ref, patterns)
-        ) {
-          labels.push(label)
-        }
-
-        return labels
-      },
-      []
-    )
-
-    if (labelsToAdd.length > 0) {
-      await tools.github.issues.addLabels({
-        number: tools.context.payload.pull_request.number,
-        labels: labelsToAdd,
-        ...repoInfo
-      })
-    }
-
-    tools.exit.success()
-  },
-  { event: 'pull_request.opened', secrets: ['GITHUB_TOKEN'] }
-)
+action()
