+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "pr-labeler",
+  "name": "pr-labeler-action",
   "private": true,
   "main": "index.js",
   "scripts": {
@@ -7,7 +7,8 @@
     "test": "jest"
   },
   "dependencies": {
-    "actions-toolkit": "^2.0.0",
+    "@actions/core": "^1.1.1",
+    "@actions/github": "^1.1.0",
     "matcher": "^2.0.0"
   },
   "devDependencies": {
